# Assignment/Chapter 03/Session 04/docs/swagger.yaml
definitions:
  request.CommentRequest:
    properties:
      message:
        type: string
      photo_id:
        type: integer
    required:
    - message
    - photo_id
    type: object
...
  response.CommentResponse:
    properties:
      created_at:
        type: string
      deleted:
        type: boolean
      deleted_at:
        type: string
      id:
        type: integer
      message:
        type: string
      photo_id:
        type: integer
      updated_at:
        type: string
      url:
        type: string
      user_id:
        type: integer
    type: object
  response.ErrorResponse:
    properties:
      error: {}
    type: object
...
info:
  contact: {}
  description: This API services is used for M-Gram Platform
  title: M-Gram Api Client
  version: "1.0"
paths:
  /v1/comment/:
    get:
      consumes:
      - application/json
      description: get all comment
      parameters:
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      - description: api-secret
        in: header
        name: api-secret
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.CommentResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            items:
              $ref: '#/definitions/response.ErrorResponse'
            type: array
      summary: get all
      tags:
      - comment
    post:
      consumes:
      - application/json
      description: add new comment
      parameters:
      - description: comment request is mandatory
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.CommentRequest'
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      - description: api-secret
        in: header
        name: api-secret
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.CommentResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            items:
              $ref: '#/definitions/response.ErrorResponse'
            type: array
      summary: add new comment
      tags:
      - comment
  /v1/comment/{id}:
    delete:
      consumes:
      - application/json
      description: delete comment by id
      parameters:
      - description: comment_id
        in: path
        name: id
        required: true
        type: integer
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      - description: api-secret
        in: header
        name: api-secret
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.CommentResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            items:
              $ref: '#/definitions/response.ErrorResponse'
            type: array
      summary: delete comment
      tags:
      - comment
    get:
      consumes:
      - application/json
      description: get comment by id
      parameters:
      - description: comment_id
        in: path
        name: id
        required: true
        type: integer
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      - description: api-secret
        in: header
        name: api-secret
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.CommentResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            items:
              $ref: '#/definitions/response.ErrorResponse'
            type: array
      summary: get comment
      tags:
      - comment
    put:
      consumes:
      - application/json
      description: update comment by id
      parameters:
      - description: comment_id
        in: path
        name: id
        required: true
        type: integer
      - description: comment request is mandatory
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.CommentRequest'
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      - description: api-secret
        in: header
        name: api-secret
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.CommentResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            items:
              $ref: '#/definitions/response.ErrorResponse'
            type: array
      summary: update comment
      tags:
      - comment
...
swagger: "2.0"